Rebase to on top of NCS v2.5.0
This points our NCS module to a branch that is v2.5.0 upstream NCS tag,
with the changes from the `gminn/lte-metrics` branch rebased on top.

I'm also updating to latest 1.5.0-rc branch for the Memfault SDK, which
has some other fixes for FOTA (namely, handle CDN cert retry loop).

And setting a few more Kconfig options:

- disable the NRF_CLOUD_FOTA module, it's pretty large and we don't use
  it
- got back to default log level of 3, INFO, which is much more useful
  when testing vs. 2=WARNING

This works for FOTA, though might be unrelated to the actual root cause.
But I think it's cleaner to base our test app on a released NCS branch.